requirements.txt
beautifulsoup4==4.3.2
blinker==1.3
Flask==0.10.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
html5lib==0.99999
itsdangerous==0.24
Jinja2==2.7.3
line-profiler==1.0
lxml==3.4.4
MarkupSafe==0.23
matplotlib==1.4.3
numpy==1.9.2
pandas==0.16.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
requests-cache==0.4.10
scikit-learn==0.16.1
scipy==0.15.1
six==1.9.0
SQLAlchemy==1.0.5
TA-Lib==0.4.9
Werkzeug==0.10.4
WTForms==2.0.2
